Conversions for 12.5 custom resources
Starting with ingredient_config and omnibus_service
