--- libraries/omnibus_service_resource.rb
@@ -1,29 +1,0 @@
-#
-# Author:: Joshua Timberman <joshua@chef.io
-# Copyright (c) 2015, Chef Software, Inc. <legal@chef.io>
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#    http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-#
-class Chef
-  class Resource
-    class OmnibusService < Chef::Resource::LWRPBase
-      resource_name :omnibus_service
-
-      actions %i(start stop restart hup int kill graceful_kill once)
-      default_action :nothing
-
-      attribute :ctl_command, kind_of: String
-      attribute :service_name, kind_of: String, regex: %r{[\w-]+\/[\w-]+}, name_attribute: true
-    end
-  end
-end
